DEV: Simplify watched word data structures
Watched words used to be returned as a hash where the key was the
word and the value was a hash with the other attributes. This data
structure did not have any advantages and made processing more
difficult.

This commit changes the data structure to be an array of hashes
where each hash contains all the attributes of the watched word.

This commit also makes it more explicit that there is a short and fast
regular expression (key "regexp") that is used to perform quick
searches and does not match the whole word, and a long and complete
regular expression (key "full_regexp") that can be used to perform
Unicode matching and may include word boundaries.