FIX: Fix category selectors for lazy loaded categories
category-drop has been fixed in a previous commit. For fixing both
category-chooser and category-selector we can use a similar pattern:
try to async load the category and notify that the property has changed
when the category is loaded, even though the value has not changed.

This pattern worked well for category-chooser because its attributes
are category IDs, but for category-selector it did not work as it
accepts a category list, which implies that the categories were loaded
already. For this reason, we introduced a new set of attributes that
use IDs instead.